refactor(admin): refactored common library
