--- packages/admin/src/components/AppHeader/AppHeader.module.css
@@ -1,8 +1,0 @@
-.menu {
-  top: 3.3rem;
-  right: 1.6rem;
-}
-
-.paths {
-  right: 10rem;
-}
