+++ packages/admin/src/pages/Logout/Logout.page.tsx
@@ -1,7 +1,7 @@
 import { Loader } from '@codement/ui';
 import React from 'react';
 
-import { Auth } from '../../containers/Auth.container';
+import { Auth } from '@codement/ui/lib/containers/Auth.container';
 
 
 export const LogoutPage = () => {
