+++ packages/ui/lib/apollo/config.ts
@@ -1,0 +1,7 @@
+export const CONFIG = {
+  localStorageTokenKey: 'token',
+  localStorageEmailKey: 'email',
+  apiHost: '%%API_HOST%%',
+  apiUrl: '/graphql',
+  prefix: ''
+};
