+++ packages/ui/lib/apollo/index.ts
@@ -1,5 +1,6 @@
 /* eslint-disable no-console, no-underscore-dangle */
 import fragments from '@codement/api/fragments.json';
+import { LocalStorage } from '@codement/ui/lib/localStorage';
 import { InMemoryCache, IntrospectionFragmentMatcher } from 'apollo-cache-inmemory';
 import { ApolloClient, ApolloError } from 'apollo-client';
 import { ApolloLink } from 'apollo-link';
@@ -8,8 +9,7 @@ import { onError } from 'apollo-link-error';
 import { createUploadLink } from 'apollo-upload-client';
 import axios from 'axios';
 
-import { CONFIG } from '../config';
-import { LocalStorage } from './localStorage';
+import { CONFIG } from './config';
 
 
 const { buildAxiosFetch } = require('@lifeomic/axios-fetch');
